fix: error 수정

+++ testDataSet/createDataset.py
@@ -168,7 +168,7 @@ def create():
             choice = np.random.choice(grade, total_lec, p=[0.1, 0.2, 0.2, 0.15, 0.15, 0.05, 0.05, 0.05,
                                                            0.05]).tolist()  # grade의 각 값이 나올 확률 지정
             dataset['등급'] = choice
-
+            dataset['분반'] = dataset['분반'].astype(int)
             dataset['분반'] = dataset['분반'].astype(str)
 
             for index, row in dataset.iterrows():
